requirements.txt
streamlit
langchain
langchain-openai
langchain-core
fastapi
uvicorn
pydantic
python-dotenv